Fix HeartRate class interface
Changed getMaximumHeartRate argument from 'float' to 'int'

+++ cpp-how-to-program/chapter-3-introducction-to-classes-objects-and-strings/3.16-mad-target-heart-rate-calculator/HeartRate.h
@@ -16,7 +16,7 @@ public:
 	void setBirthYear(int);
 	int getBirthYear();
 	int getAge(); // Calculate the age
-	float getMaximumHeartRate(float); // Calculate maximum heart rate
+	float getMaximumHeartRate(int); // Calculate maximum heart rate
 	float getTargetHeartRate(float); // Calculate target heart rate
 private:
 	std::string firstName;
